pyproject.toml
[project]
name = "rag-custom"
version = "0.1.0"
description = "Rag custom app."
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "numpy<2",
    "torch<2.3",
    "unstructured[all-docs]>=0.18.27",
    "sentence-transformers==5.2.0",
    "chromadb>=1.4.1",
    "langchain>=1.2.6",
    "langchain-community>=0.4.1",
    "langchain-ollama==1.0.1",
    "lxml>=6.0.2",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
]
